feat(subcluster): add support for labelled subclusters and paramterized lxd_nodes_per_subcluster
This commit exposes a variable called `lxd_nodes_per_subcluster` to
change the number of lxd nodes per subcluster.

This commit also changes the schema for subclusters per region from
accepting just a number to a list of strings which will allow a user to
customize the name for each subcluster and enable the user to do
targeted removal of a specific subcluster in the future.